fixed la dernière requete E N F I N. Ajout des tests manquant dans StudentTest.java, début sur les autres, je finis au moins subject ce soir

+++ src/main/java/fr/uga/im2ag/l3/miage/db/repository/impl/StudentRepositoryImpl.java
@@ -52,7 +52,7 @@ public class StudentRepositoryImpl extends BaseRepositoryImpl implements Student
     public List<Student> findStudentHavingGradeAverageAbove(float minAverage) {
     	List<Student> Res;
         Res = entityManager.createNamedQuery("Student-AvgAbove", Student.class)
-        		.setParameter("avg", minAverage)
+        		.setParameter("avg", (double)minAverage)
         		.getResultList();
         return Res;
     }
